requirements-test.txt
# Test Framework
pytest==8.0.0
pytest-cov==4.1.0
pytest-xdist==3.5.0  # For parallel test execution
pytest-timeout==2.2.0  # For test timeouts
pytest-benchmark==4.0.0  # For performance benchmarking

# Mock and Fixtures
pytest-mock==3.12.0
pytest-fixture-config==1.7.0
pytest-env==1.1.3

# Test Utilities
coverage==7.4.1
psutil==5.9.8  # For resource monitoring
hypothesis==6.98.0  # For property-based testing
freezegun==1.4.0  # For time-based testing

# Video Processing
opencv-python==4.9.0.80
moviepy==1.0.3
ffmpeg-python==0.2.0

# AI Models (Test Versions)
openai==1.12.0
google-ai-generativelanguage==0.4.0
twelvelabs==0.4.4  # Using latest version for API v1.3 compatibility

# Vector Storage
numpy==1.26.3
faiss-cpu==1.7.4  # CPU version for testing

# Utilities
python-dotenv==1.0.1
requests-mock==1.11.0
responses==0.24.1  # For mocking HTTP requests
tenacity==8.2.3  # For retrying operations
